refactor: clear code
- 코드 정리
- 인수인계 준비

obab-client #8
